# updatePokemonDom+ searchPokemon - functions
1) updatePokemonDom-
The function is based on the object "PokemonObject" and based on the information in which it updates the relevant DOM for the selected Pokemon.

2) createTypesList-
The function gets an array with string types. For each type in the array creates an option type object and makes a append  child to the relevant object.

3) cleanTypesList-
The function takes all option-type objects and deletes them.

4) searchPokemon -
The function is a handler for clicking the Search button. It sends a GET request to the API and receives Pokemon information based on the search input. It then updates the object "PokemonObject" and calls the function "updatePokemonDom" in order to update the DOM.