# Create name lists by type selection
1)getTypeUrl -
When you select a particular type in the list of types area, the function will locate its index in the "PokemonObject" object  in the key "typeList" and access the url in that "PokemonObject" object  in the key "namesRelatedToTypesUrls" using the same index.
It will then send it to a function "getType" that will search for it in the API.

2) getType -
The function receives a URL and sends a request to the API to get information about the same TYPE associated with the URL.
From the information it builds an array with the names of all the Pokemon that contain the same type and calls the function "NameListToDOM" with the same array as an argument.

3) NameListToDOM -
The function gets an array of Pokemon names, it builds within the list names of option type elements that contain the name of the Pokemon in the array.

4) cleanNamesList -
The function deletes the option elements that contain a Pokemon name in the list of names.