# Loader
When sending a request to the API, the loader will be activated and stopped when the answer or error is returned.

1) playLoader -
The function creates an image type element. The element gets the class loader by which it gets the design and animation. An element added to the DOM as a child of the search area.

2) stopLoader -
The function removes the element of the Loader from the DOM