# Search next and previous buttons
1)movePokemon -
The function checks which of the buttons is pressed and uses the object ... in order to locate the ID of the next Pokemon or of the previous Pokemon.
If we click forward at the end of the list we will reach the first Pokemon and so on in the opposite direction.